Remove SGD2MAPI ext. glide3x D2DX C++98 hpp file
