+++ SlashGaming-Diablo-II-Free-Resolution/src/patches/required/d2glide_set_display_width_and_height_patch/d2glide_set_display_width_and_height.cc
@@ -48,7 +48,6 @@
 #include <sgd2mapi.hpp>
 
 #include "../../../helper/game_resolution.hpp"
-#include "../../../helper/glide3x_d2dx.hpp"
 #include "../../../sgd2mapi_extension/sgd2mapi_extension.hpp"
 
 namespace sgd2fr::patches {
@@ -87,7 +86,7 @@ void __cdecl Sgd2fr_D2Glide_SetDisplayWidthAndHeight(
 
   if (::d2::glide3x_library_version::GetRunning()
       == ::d2::glide3x_library_version::kD2dx) {
-    d2dx_glide::SetCustomResolution(*width, *height);
+    SetCustomResolution(*width, *height);
   }
 }
