+++ SlashGaming-Diablo-II-Free-Resolution/src/sgd2mapi_extension/glide3x_library_d2dx/glide3x_library_d2dx.h
@@ -43,8 +43,8 @@
  *  work.
  */
 
-#ifndef SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_D2DX_GLIDE3X_D2DX_H_
-#define SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_D2DX_GLIDE3X_D2DX_H_
+#ifndef SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_LIBRARY_D2DX_GLIDE3X_LIBRARY_D2DX_H_
+#define SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_LIBRARY_D2DX_GLIDE3X_LIBRARY_D2DX_H_
 
 #include <windows.h>
 
@@ -70,4 +70,4 @@ HRESULT GetSuggestedCustomResolution(int* width, int* height);
 } /* extern "C" */
 #endif /* __cplusplus */
 
-#endif /* SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_D2DX_GLIDE3X_D2DX_H_ */
+#endif /* SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_LIBRARY_D2DX_GLIDE3X_LIBRARY_D2DX_H_ */
